Assignements/Assignment-1/requirement.txt
streamlit
python-dotenv
pydantic
langchain-groq
langchain-core